// src/lib/api/models/user/user.ts
import { components, paths } from "@lib/api/owaru"; // <- generated by openapi-typescript
import tt from "@public/locales/es/common.json"; //just the key

export type UserSession = {
	token_data?: API_Token;
	user?: User;
}

export type User = UserDataOptional;


export type API_Token = Token;

type Token = components["schemas"]["Token"];
type UserData = components["schemas"]["UserInput"] & { id: number } ;

// user data with optional fields
type UserDataOptional = Partial<UserData>;

export type TokenDecoded = {
	exp?: number;
	sub?: string;
	company_id: string;
};

export enum Roles {
	ADMIN = 1,
	STUDENT = 2,
}